requirements.txt
ipwhois
requests
python-whois
hyper
beautifulsoup4
pandas
aiohttp
